(local) slice + dice.
