+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-shell-authorizer/src/main/java/org/apache/nifi/authorization/ShellCommandsProvider.java
@@ -1,0 +1,9 @@
+package org.apache.nifi.authorization;
+
+interface ShellCommandsProvider {
+    String getUsersList();
+    String getUserGroups();
+    String getGroupsList();
+    String getGroupMembers();
+    String getSystemCheck();
+}
