+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-shell-authorizer/src/test/java/org/apache/nifi/authorization/ShellUserGroupProviderTest.java
@@ -17,30 +17,17 @@
 package org.apache.nifi.authorization;
 
 import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.net.URL;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 
 
-import java.time.Duration;
 import java.util.Set;
 
-import org.apache.commons.io.FileUtils;
 import org.junit.Before;
-import org.junit.Ignore;
-import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
 
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -51,87 +38,27 @@ import static org.mockito.Mockito.mock;
 import org.junit.rules.TemporaryFolder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.testcontainers.containers.BindMode;
+
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.utility.MountableFile;
 
 
-abstract class ShellUserGroupProviderBase {
-    private final String KNOWN_USER  = "root";
-    private final String KNOWN_UID   = "0";
-
-    private final String KNOWN_GROUP = "root";
-    private final String OTHER_GROUP = "wheel"; // e.g., macos
-    private final String KNOWN_GID   = "0";
-    
-    public void testGetUsers(UserGroupProvider provider) throws Exception {
-        Set<User> users = provider.getUsers();
-        assertNotNull(users);
-        assertTrue(users.size() > 0);
-    }
-
-    public void testGetUser(UserGroupProvider provider) throws Exception {
-        User root = provider.getUser(KNOWN_UID);
-        assertNotNull(root);
-        assertEquals(KNOWN_USER, root.getIdentity());
-        assertEquals(KNOWN_UID, root.getIdentifier());
-    } 
-   
-    public void testGetUserByIdentity(UserGroupProvider provider) throws Exception {
-        User root = provider.getUserByIdentity(KNOWN_USER);
-        assertNotNull(root);
-        assertEquals(KNOWN_USER, root.getIdentity());
-        assertEquals(KNOWN_UID, root.getIdentifier());
-    }
-
-    public void testGetGroups(UserGroupProvider provider) throws Exception {
-        Set<Group> groups = provider.getGroups();
-        assertNotNull(groups);
-        assertTrue(groups.size() > 0);
-    }
-
-    public void testGetGroup(UserGroupProvider provider) throws Exception {
-        Group group = provider.getGroup(KNOWN_GID);
-        assertNotNull(group);
-        assertTrue(group.getName().equals(KNOWN_GROUP) || group.getName().equals(OTHER_GROUP));
-        assertEquals(KNOWN_GID, group.getIdentifier());
-    }
-
-    public void testGroupMembership(UserGroupProvider provider) throws Exception {
-        Group group = provider.getGroup(KNOWN_GID);
-        assertNotNull(group);
-        assertTrue(group.getUsers().size() > 0);
-        assertTrue(group.getUsers().contains(KNOWN_USER));
-    }
-
-    public void testGetUserAndGroups(UserGroupProvider provider) throws Exception {
-        UserAndGroups user = provider.getUserAndGroups(KNOWN_UID);
-        assertNotNull(user);
-        assertTrue(user.getGroups().size() > 0);
-
-        Set<Group> groups = provider.getGroups();
-        assertTrue(groups.size() > user.getGroups().size());
-    }
-}
-
-
 public class ShellUserGroupProviderTest extends ShellUserGroupProviderBase {
-    public static final Logger logger = LoggerFactory.getLogger(ShellUserGroupProviderTest.class);
+    private static final Logger logger = LoggerFactory.getLogger(ShellUserGroupProviderTest.class);
     
-    public final static String ALPINE_IMAGE = "panubo/sshd:latest";
-    public final static String CENTOS_IMAGE = "natural/centos-sshd:latest";
-    public final static String DEBIAN_IMAGE = "natural/debian-sshd:latest";
-    public final static String UBUNTU_IMAGE = "natural/ubuntu-sshd:latest";
-    public final static List<String> TEST_CONTAINER_IMAGES = Arrays.asList(ALPINE_IMAGE
+    private final static String ALPINE_IMAGE = "panubo/sshd:latest";
+    private final static String CENTOS_IMAGE = "natural/centos-sshd:latest";
+    private final static String DEBIAN_IMAGE = "natural/debian-sshd:latest";
+    private final static String UBUNTU_IMAGE = "natural/ubuntu-sshd:latest";
+    private final static List<String> TEST_CONTAINER_IMAGES = Arrays.asList(ALPINE_IMAGE
                                                                            , CENTOS_IMAGE
                                                                            , DEBIAN_IMAGE
                                                                            , UBUNTU_IMAGE);
     
-    public final static String CONTAINER_SSH_AUTH_KEYS = "/root/.ssh/authorized_keys";
-    public final static Integer CONTAINER_SSH_PORT = 22;
+    private final static String CONTAINER_SSH_AUTH_KEYS = "/root/.ssh/authorized_keys";
+    private final static Integer CONTAINER_SSH_PORT = 22;
     
     private AuthorizerConfigurationContext authContext;
-    private ClassLoader classLoader;
     private ShellUserGroupProvider localProvider;
     private String sshPrivKeyFile;
     private String sshPubKeyFile;
@@ -143,8 +70,6 @@ public class ShellUserGroupProviderTest extends ShellUserGroupProviderBase {
     
     @Before
     public void setup() throws IOException {
-        classLoader = Thread.currentThread().getContextClassLoader();
-        
         authContext = mock(AuthorizerConfigurationContext.class);
         initContext = mock(UserGroupProviderInitializationContext.class);
 
@@ -166,42 +91,43 @@ public class ShellUserGroupProviderTest extends ShellUserGroupProviderBase {
     }
 
     @Test
-    public void testGetUsers() throws Exception {
+    public void testGetUsers() {
         testGetUsers(localProvider);
     }
 
     
     @Test
-    public void testGetUser() throws Exception {
+    public void testGetUser() {
         testGetUser(localProvider);
     }
     
     @Test
-    public void testGetUserByIdentity() throws Exception {
+    public void testGetUserByIdentity() {
         testGetUserByIdentity(localProvider);
     }
 
     @Test
-    public void testGetGroups() throws Exception {
+    public void testGetGroups() {
         testGetGroups(localProvider);
     }
     
     @Test
-    public void testGetGroup() throws Exception {
+    public void testGetGroup() {
         testGetGroup(localProvider);
     }    
 
     @Test
-    public void testGroupMembership() throws Exception {
+    public void testGroupMembership() {
         testGroupMembership(localProvider);
     }
         
     @Test
-    public void testGetUserAndGroups() throws Exception {
+    public void testGetUserAndGroups() {
         testGetUserAndGroups(localProvider);
     }
         
-    public GenericContainer createContainer(String image) throws IOException, InterruptedException {
+    @SuppressWarnings("RedundantThrows")
+    private GenericContainer createContainer(String image) throws IOException, InterruptedException {
         GenericContainer container = new GenericContainer(image)
             .withEnv("SSH_ENABLE_ROOT", "true")
             .withExposedPorts(CONTAINER_SSH_PORT);
@@ -214,7 +140,7 @@ public class ShellUserGroupProviderTest extends ShellUserGroupProviderBase {
         return container;
     }
     
-    public UserGroupProvider createRemoteProvider(GenericContainer container) {
+    private UserGroupProvider createRemoteProvider(GenericContainer container) {
         final ShellCommandsProvider remoteCommands = RemoteShellCommands.wrapOtherProvider(new NssShellCommands(),
                                                                                            sshPrivKeyFile,
                                                                                            container.getContainerIpAddress(),
@@ -228,7 +154,7 @@ public class ShellUserGroupProviderTest extends ShellUserGroupProviderBase {
     }
     
     @Test
-    public void testVariousSystemImages() throws Exception {
+    public void testVariousSystemImages() {
         TEST_CONTAINER_IMAGES.forEach(image -> {
                 GenericContainer container;
