+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/pom.xml
@@ -47,6 +47,7 @@
         <module>nifi-properties-loader</module>
         <module>nifi-standard-prioritizers</module>
         <module>nifi-mock-authorizer</module>
+        <module>nifi-shell-authorizer</module>
     </modules>
     <dependencies>
         <dependency>
